+++ expectise/__init__.py
@@ -1,6 +1,6 @@
-from .expect import Expect
-from .expectations import Expectations
-from .mocks import disable_mock
-from .mocks import mock
-from .mocks import mock_if
-from .mocks import tear_down
+from .hooks import disable_mock
+from .hooks import mock
+from .hooks import mock_if
+from .hooks import tear_down
+from .mock.expect import Expect
+from .mock.expectations import Expectations
