+++ expectise/hooks/tear_down.py
@@ -1,0 +1,6 @@
+from expectise.mock.session import session
+
+
+def tear_down():
+    """Reset mocking behavior so that further tests can be run without any interference."""
+    session.tear_down()
